notebook_CLI/backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
python-dotenv==1.2.2
motor==3.3.2
pymysql==1.1.0
SQLAlchemy==2.0.49
python-multipart==0.0.9
groq==0.37.1
langchain==1.2.15
langchain-groq==1.1.2
langchain-huggingface==1.2.2
langchain-community==0.4.1
langchain-core==1.3.0
langchain-text-splitters==1.1.2
sentence-transformers==5.4.1
qdrant-client==1.9.1
pypdf==6.10.2
huggingface_hub==1.11.0
torch==2.11.0
transformers==5.5.4
tokenizers==0.22.2
numpy==2.4.4
httpx==0.28.1
httpx-sse==0.4.3
pydantic==2.13.2
